eye_tracker_fixation_scroll

corrects the eye-tracking coordinate data to fit in a webpage scrolled vertically by the participant

Usage

eye_tracker_fixation_scroll(
  eyes_data,
  timestamp_start,
  timestamp_stop,
  image_height,
  image_width,
  calibration,
  time_shift = 0,
  starting_scroll = 0,
  output_file = "eye_tracker_fixation_corrected.csv",
  anchors = list(),
  rules = list(),
  outside_image_is_na = TRUE,
  na.rm = TRUE
)

Arguments

eyes_data

The dataset

timestamp_start

The starting timestamp from the .csv file at which the participant was watching the scrollable (AFTER time_shift applied)

timestamp_stop

The final timestamp from the .csv file at which the participant was watching the scrollable (AFTER time_shift applied)

image_height

The total height of the image (in pixels)

image_width

The total width of the image (in pixels)

calibration

A calibration list (see the scroll_calibration function)

time_shift

[Optional] A time shift parameter to synchronize the .csv file with a video recording (e.g. if the recording started after the start of the .csv). Default: 0

starting_scroll

[Optional] If the participant did not start watching the webpage from the top, you can indicate the y coordinate at which he started here (in pixels). Default: 0

output_file

[Optional] The name of the output .csv file. If an empty string, will just return the data without creating a new file. Default: empty string

anchors

[Optional] A list of potentially immovable areas inside the scrollable (e.g. anchored menus in a website) - see manual for a correct usage. Default: empty list

rules

[Optional] A list of functions that can act as rules to activate/deactivate immovables areas in the scrollable (e.g. a menu that disappears after X pixels have been scrolled) - See manual for a correct. Default: empty list

outside_image_is_na

[Optional] Indicates if values outside the AOI (e.g. the windows bar) should be set to NA or kept in the file/dataset. If set to FALSE, coordinates above/before the AOI will become negative. Default: TRUE

na.rm

[Optional] Indicates if lines with NA y values should be dropped in the final file/dataset. Default: TRUE

Value

The Dataset with corrected x and y's

Examples

if (FALSE) {
test_anchors <- list(array(c(c(0,0,1902,95), c(0,0,1902,95),
    c(1607,96,1902,905), c(1607,96,1902,905)), dim = c(4,2,2)),
    array(c(c(0,0,1902,63), c(0,32,1902,95), c(1607,63,1902,900),
    c(1607,63,1902,900)), dim = c(4,2,2)), array(c(c(1607,906,1920,1080),
    c(1607,6481,1920,6655)), dim = c(4,2,1)))
test_rules = list(rule_before_scrolling, rule_after_scrolling, rule_true)
test_calibration <- scroll_calibration(1920, 1080, 88, 0, 40, 0, 100)
test_data <- eye_tracker_fixation_scroll(anchors = test_anchors,
    rules = test_rules, calibration = test_calibration)
}